extract Conter component
