+++ src/Screens/ReactGym/styledReactGym.js
@@ -13,31 +13,3 @@ export const GYMContainer = styled.div`
   justify-content: center;
   align-items: center;
 `;
-
-export const CounterContainer = styled.div`
-  display: grid;
-  place-items: center;
-`;
-
-export const CounterText = styled.p`
-  color: #fff;
-  font-weight: 700;
-  font-size: 4rem;
-  margin: 0;
-`;
-
-export const CounterButtons = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0rem 5rem;
-`;
-
-export const CounterButton = styled.button`
-  font-size: 2rem;
-  color: ${({ theme }) => theme.colors.titleColor};
-  background: ${({ theme }) => theme.colors.scrollBarColor};
-  padding: 1rem 2rem;
-  border-radius: 0.4rem;
-  transition: all 0.4s ease;
-`;
